Add files via upload
See, what happened was I had the temerity to upload the file to Github in a repository, but then I edited said file in Jupyter. Of course, Github Desktop didn't recognize it and wouldn't accept anything approaching a new commit. So I just re-uploaded the whole thing. Total cluster. Oh well! I'm going to delete this as soon as it is recognized.